Radius new criteria

+++ src/neurons.hh
@@ -22,7 +22,7 @@ namespace Geometry
             int width_;
             int height_;
 
-      double getRadius(int i);
+      double getRadius(double i);
     };
 }
